requirements.txt
PyYAML==6.0.2
garth==0.5.2
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv
langchain-core
langchain-openai
langgraph
pydantic>=2.0.0
requests
garminconnect
openai

# Database dependencies
sqlalchemy>=2.0.0
alembic>=1.12.0